requirements.txt
GitPython
PyYAML

